Update extract_capture.pl
Update: You can now specify a directory that doesn't contain datadir folders as the input directory. If the required index and .pic files are found, they will be processed. Note that if you have multiple datadir folders, the input directory should be the directory that contains the datadir folders so that all of them are processed.

Fixed: Processing order of datadir folders when the script runs on Linux.